+++ app/src/main/AndroidManifest.xml
@@ -14,8 +14,19 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ShyPolarBear"
         tools:targetApi="31">
+
+<!--        <activity-->
+<!--            android:name="com.shypolarbear.presentation.MainActivity"-->
+<!--            android:exported="true"-->
+<!--            android:label="@string/app_name"-->
+<!--            android:theme="@style/Theme.ShyPolarBear">-->
+<!--            <intent-filter>-->
+<!--                <action android:name="android.intent.action.MAIN"/>-->
+<!--                <category android:name="android.intent.category.LAUNCHER"/>-->
+<!--            </intent-filter>-->
+<!--        </activity>-->
         <activity
-            android:name="com.shypolarbear.presentation.MainActivity"
+            android:name="com.shypolarbear.presentation.sample.SampleActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ShyPolarBear">
@@ -24,6 +35,7 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+
     </application>
 </manifest>
