+++ presentation/src/main/java/com/shypolarbear/presentation/sample/SampleActivity.kt
@@ -1,0 +1,20 @@
+package com.shypolarbear.presentation.sample
+
+import androidx.activity.viewModels
+import com.shypolarbear.presentation.R
+import com.shypolarbear.presentation.base.BaseActivity
+import com.shypolarbear.presentation.databinding.ActivitySampleBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class SampleActivity : BaseActivity<ActivitySampleBinding, SampleViewModel>(
+    R.layout.activity_sample
+) {
+    override val viewModel: SampleViewModel by viewModels()
+
+    override fun initView() {
+        binding.apply {
+
+        }
+    }
+}
